requirements.txt
django==4.2.1
djangorestframework==3.15.2
drf-yasg==1.21.7
python-dotenv==1.0.1
redis==5.1.1
hiredis==3.0.0
django-debug-toolbar==4.2.0